+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - added action to reload extension in case of error
 - added possibility to edit models coming from extension provider
 - prevent crash from circular references in error logging
+- added support for using custom and extension commands in saved prompts
 
 ## [0.75.0]